+++ CHANGELOG.md
@@ -1,4 +1,5 @@
 ### v0.2.14
+- New config option: AutoPickupExcludeFodder: exclude food items for tames from auto-pickup if tames are within search range (defaults to true)
 - Fix Auto-Collection Stops After Instant Resource Consumption [#32](https://github.com/ArgusMagnus/ValheimServersideQoL/issues/32)
 - Tweaks and bug fixes